requirements.txt
nltk
python-docx
punctuator
rpunct

